精简 jy901 驱动; 控制增加死区

+++ user/data-type/propeller.h
@@ -9,11 +9,11 @@ struct propeller_parameters
     uint8_t enabled : 1;
     uint8_t reversed : 1;
     int channel;
-    double deadzone_upper;
-    double deadzone_lower;
+    int deadzone_upper;  // 0 ~ 1000
+    int deadzone_lower;  // -0 ~ -1000
     double power_negative;
     double power_positive;
-    double power_cur;
+    double power_cur;   // -1.0 ~ 1.0
     double power_last;
 };
